Moved Udemy's entities
